reorganize authentication logic
