+++ src/RemoteFlowsProvider.tsx
@@ -1,15 +1,9 @@
-import { createClient } from '@hey-api/client-fetch';
-import {
-  QueryClient,
-  QueryClientProvider,
-  useQuery,
-} from '@tanstack/react-query';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import type { PropsWithChildren } from 'react';
-import React, { useRef } from 'react';
+import React from 'react';
 
-import { ENVIROMENTS } from '@/src/environments';
 import { ThemeProvider } from '@/src/theme';
-import { client } from './client/client.gen';
+import { useAuth } from './auth';
 import { FormFieldsContext, RemoteFlowContext } from './context';
 import { Components, RemoteFlowsSDKProps } from './types/remoteFlows';
 
@@ -26,40 +20,12 @@ function RemoteFlowContextWrapper({
   auth,
   isTestingMode,
 }: RemoteFlowContextWrapperProps) {
-  const session = useRef<{ accessToken: string; expiresAt: number } | null>(
-    null,
-  );
-  const { refetch } = useQuery({
-    queryKey: ['auth'],
-    queryFn: auth,
-    enabled: false,
+  const remoteApiClient = useAuth({
+    auth,
+    options: {
+      isTestingMode: !!isTestingMode,
+    },
   });
-
-  const baseUrl = isTestingMode
-    ? ENVIROMENTS.partners
-    : process.env.REMOTE_GATEWAY_URL;
-
-  const remoteApiClient = useRef(
-    createClient({
-      ...client.getConfig(),
-      baseUrl,
-      auth: async () => {
-        function hasTokenExpired(expiresAt: number | undefined) {
-          return !expiresAt || Date.now() + 60000 > expiresAt;
-        }
-        if (!session.current || hasTokenExpired(session.current.expiresAt)) {
-          const { data } = await refetch();
-          if (data) {
-            session.current = {
-              accessToken: data.accessToken,
-              expiresAt: Date.now() + data.expiresIn * 1000,
-            };
-          }
-        }
-        return session.current?.accessToken;
-      },
-    }),
-  );
   return (
     <RemoteFlowContext.Provider value={{ client: remoteApiClient.current }}>
       {children}
