+++ src/environments.ts
@@ -1,4 +1,4 @@
-export const ENVIROMENTS = {
+export const ENVIRONMENTS = {
   staging: 'https://gateway.niceremote.com',
   partners: 'https://gateway.partners.remote-sandbox.com',
   production: 'https://gateway.remote.com',
